+++ CHANGELOG.md
@@ -6,11 +6,16 @@ The Product Changelog at **[matomo.org/changelog](https://matomo.org/changelog)*
 
 ## Matomo 3.9.0
 
+### Breaking Changes
+
+* `Referrers.getKeywordsForPageUrl` and `Referrers.getKeywordsForPageTitle` APIs have been deprecated and will be removed in Matomo 4.0.0
+
 ### New Features
 * It is now possible to use monolog's FingersCrossedHandler which buffers all logs and logs all of them in case of warning or error.
 
 ### New APIs
 * A new tracker method `disableQueueRequest` has been added to disable queued requests which may be useful when logs are imported.
+* The event `LanguageManager.getAvailableLanguages` has been deprecated. Use `LanguagesManager.getAvailableLanguages` instead.
 
 ## Matomo 3.8.0